pyproject.toml
[project]
name = "csvtopdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fpdf2>=2.7",
]
